moving the application editing modal window and fix date format
